Розділив схеми,виправив помилки
