+++ routes/api/contacts.js
@@ -4,7 +4,7 @@ const ctrl = require("../../controllers/contacts");
 
 const { validateBody, isValidId } = require("../../middlewares");
 
-const { schemas } = require("../../utils/validation/contactValidationSchemas");
+const { schemas } = require("../../utils/validation");
 
 const router = express.Router();
 
@@ -12,7 +12,7 @@ router.get("/", ctrl.getAll);
 
 router.get("/:contactId", isValidId, ctrl.getById);
 
-router.post("/", isValidId, validateBody(schemas.addContactSchema), ctrl.add);
+router.post("/", validateBody(schemas.addContactSchema), ctrl.add);
 
 router.put(
   "/:contactId",
