+++ utils/validation/contactValidationSchemas.js
@@ -1,30 +1,4 @@
-const { Schema, model } = require("mongoose");
-
 const Joi = require("joi");
-
-const { handleMongooseError } = require("../../helpers");
-
-const contactSchema = new Schema(
-  {
-    name: {
-      type: String,
-      required: [true, "Set name for contact"],
-    },
-    email: {
-      type: String,
-    },
-    phone: {
-      type: String,
-    },
-    favorite: {
-      type: Boolean,
-      default: false,
-    },
-  },
-  { versionkey: false, timestamps: false }
-);
-
-contactSchema.post("save", handleMongooseError);
 
 const addContactSchema = Joi.object({
   name: Joi.string().required(),
@@ -33,8 +7,8 @@ const addContactSchema = Joi.object({
   favorite: Joi.boolean(),
 });
 const updateStatusContactSchema = Joi.object({
-  fovorite: Joi.boolean(),
+  favorite: Joi.boolean().required(),
 });
 const schemas = { addContactSchema, updateStatusContactSchema };
-const Contact = model("contact", contactSchema);
-module.exports = { Contact, schemas };
+
+module.exports = schemas;
