+++ utils/validation/index.js
@@ -1,0 +1,3 @@
+const schemas = require("./contactValidationSchemas");
+const Contact = require("./mongooseValidationSchema");
+module.exports = { Contact, schemas };
