+++ utils/validation/mongooseValidationSchema.js
@@ -1,0 +1,26 @@
+const { Schema, model } = require("mongoose");
+const { handleMongooseError } = require("../../helpers");
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+    },
+    phone: {
+      type: String,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  { versionKey: false, timestamps: false }
+);
+
+contactSchema.post("save", handleMongooseError);
+const Contact = model("contact", contactSchema);
+
+module.exports =  Contact ;
